requirements.txt
pandas~=1.2.1
python-dotenv~=0.15.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gcsfs~=2022.5.0